refactor(sync): introduce BaseSyncRoom abstract class
WorkspaceRoom and DocumentRoom shared identical fetch, sync, getDoc, and
WebSocket lifecycle methods. Extracted into BaseSyncRoom with two hooks:
createDoc() for Y.Doc options and initRoom(sql) for extra DDL/setup.

WorkspaceRoom is now an empty subclass (12 lines). DocumentRoom retains
only its gc:false override, snapshots table, and snapshot RPC methods.
